requirements.txt
Appium-Python-Client==1.0.1
comtypes==1.1.7
selenium==3.141.0
typing==3.7.4.1
uiautomation==2.0.4
urllib3==1.25.9
